Solve #8: convert Rusle to python arcgis toolbox
A python arcgis toolbox is a pure python toolbox without any binary
file.
The structure of such a tool is different from an atbx toolbox.
The toolbox is located in a folder esri/toolboxes/.

This new structure permits to better integrate the toolbox in Arcgis.
Thus, the toolbox can be installed with pip and will appear as a
regular toolbox in the geoprocessing tools.